requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-filter==22.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
python-decouple==3.6
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sendgrid==6.9.7
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
